+++ CHANGELOG.md
@@ -10,11 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- LXD images explorer.
+- LXD networks explorer.
+- LXD storage explorer.
 - Instance explorer displays instance status
 - Instance explorer is refreshed every 15 seconds.
 - Refresh intervall can be configured via the `lxd.refreshInterval` property.
-- LXD images explorer.
-- LXD networks explorer.
 
 ### Fixed